app/src/main/java/com/example/rachelhutchison/guessinggame/model/FanDuelPlayers.java
package com.example.rachelhutchison.guessinggame.model;

import java.util.List;

public class FanDuelPlayers {
    private Meta meta;
    private List<Fixture> fixtures;
    private List<Player> players;
    private List<Team> teams;

    public Meta getMeta() {
        return meta;
    }

    public void setMeta(Meta meta) {
        this.meta = meta;
    }

    public List<Fixture> getFixtures() {
        return fixtures;
    }

    public void setFixtures(List<Fixture> fixtures) {
        this.fixtures = fixtures;
    }

    public List<Player> getPlayers() {
        return players;
    }

    public void setPlayers(List<Player> players) {
        this.players = players;
    }

    public List<Team> getTeams() {
        return teams;
    }

    public void setTeams(List<Team> teams) {
        this.teams = teams;
    }
}
